Obtenir les résultats d'une ACM pour chaque modalités.

Cette fonction permet de renvoyer un

Usage

extraire_modalites(resultat_acm)

Arguments

resultat_acm

Une résultat d'ACM sortie de FactoMineR::MCA() ou ACM().

data_acm

le tableau de donnée d'origine

Value

Un tableau avec pour chaque modalité les coordonnées, les contributions, les cos2 et les vtest pour les 5 premières dimensions.

Dans le détail, le tableau comporte les variables :

type : variables actives ou supplémentaires ?

variable : Le nom de la variable.

modalites : la modalité en question.

n : Le nombre d'individus au sein du tableau de données qui partagent cette modalité.

pourcentage : Par des individus au sein du tableau de données qui partagent cette modalité

dim1_coord, dim2_coord, ... : Coordonnée de la modalité sur l'axe 1, 2, ...

dim1_contrib, dim2_contrib,... : Contribution de chaque modalité à la contruction des axes sur l'axe 1, 2, ...

dim1_cos2, dim2_cos2, ... : Cosinus carré des modalités sur l'axe 1, 2,...

dim1_vtest, dim2_vtest, ... : Vtest des coordonnées de la modalité sur l'axe 1, 2,...

Examples

# on commence par créer un résultat d'ACM :
resultat <- createurs_de_contenus_pol |>
  ACM(variables_supplementaires = personne)
#> 
#> ── Variables actives 
#> L'ACM a été réalisé sur ces variables actives : `titre`, `orientation`,
#> `orientation2`, `debut`, `visage_decouvert`, `abonnes`, `genre`, `age`,
#> `streamer_plein_temps`, `collectif`, `livre`, `podcast`, `twitter`, `facebook`,
#> `instagram`, `tiktok`, `discord`, `twitch`, …, `partenariat_entreprise`, and
#> `autres_services`.
#> 
#> ── Variables supplémentaires 
#> Les variables supplémentaires sont les suivantes : `personne`.

# Puis on peut en extraire le tableau descriptif de chaque modalité :
resultat |> extraire_modalites()
#> # A tibble: 189 × 21
#>    type    variables modalites     n pourcentage dim1_coord dim1_cos2 dim1_vtest
#>    <chr>   <chr>     <chr>     <int>       <dbl>      <dbl>     <dbl>      <dbl>
#>  1 variab… personne  Antoine …     1         1.8      -0.53      0.01      -0.53
#>  2 variab… titre     Antoine …     1         1.8      -0.53      0.01      -0.53
#>  3 variab… titre     BLAST, L…     1         1.8      -0.86      0.01      -0.86
#>  4 variab… titre     Bolchege…     1         1.8      -0.23      0         -0.23
#>  5 variab… titre     Canard R…     1         1.8      -0.34      0         -0.34
#>  6 variab… titre     Cass And…     1         1.8      -1.14      0.02      -1.14
#>  7 variab… titre     Cercle A…     1         1.8       0.7       0.01       0.7 
#>  8 variab… personne  Christop…     1         1.8       1.19      0.03       1.19
#>  9 variab… titre     Christop…     1         1.8       1.19      0.03       1.19
#> 10 variab… personne  Clément …     1         1.8      -0.25      0         -0.25
#> # ℹ 179 more rows
#> # ℹ 13 more variables: dim1_contrib <dbl>, dim2_coord <dbl>, dim2_cos2 <dbl>,
#> #   dim2_vtest <dbl>, dim2_contrib <dbl>, dim3_coord <dbl>, dim3_cos2 <dbl>,
#> #   dim3_vtest <dbl>, dim3_contrib <dbl>, dim4_coord <dbl>, dim4_cos2 <dbl>,
#> #   dim4_vtest <dbl>, dim4_contrib <dbl>